paycompass-v2/apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.11.0
python-dotenv==1.0.1
python-multipart==0.0.20
pandas==2.2.0
chardet==5.2.0
numpy>=1.24.0
scikit-learn>=1.5.0
openai>=1.0.0
PyJWT>=2.8.0
cryptography>=42.0.0